+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 - `Element::Capabilities::Auditable::Taint`
     - Added longest-word-optimization -- Checks if the longest word of a regexp
         exists in the response body prior to matching the full-blown regexp.
+- `Element::Cookie#encode`
+    - Allow `=` to remain un-encoded in the cookie value.
 - `Options#load` -- Updated to support serialized `Hash` objects.
 - Modules
     - Audit